Fix compile warningh

+++ src/include/string.h
@@ -4,7 +4,7 @@
 #include <type.h>
 #include <peripheral/uart.h>
 void* memcpy(void* dest, const void* src, size_t n);
-void* memset(void* dest, uint8_t val, size_t n);
+void* memset(void* dest, int val, size_t n);
 
 size_t strlen(const char* s);
 int strcmp(const char* s1, const char* s2);
